refactor(n1): add server-read storefront seam for initial page data
